backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.2
pymongo==4.6.0
pydantic==2.8.2
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
email-validator==2.1.0
groq==0.30.0
python-dotenv==1.0.0
